+++ CHANGELOG.rst
@@ -2,6 +2,10 @@
 Changelog for package rodney_sim_control
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
+Unreleased
+------------------
+* Changes so that simulation can run mission 1
+
 0.6.0 (2019-04-18)
 ------------------
 * Added configuration and launching of the nav stack